SentimentModeling v2.2
